+++ ChangeLog.md
@@ -2,6 +2,8 @@
 
 * **24.09.18**:
     - Added server passwords within login modal
+    - Fixed native encoding for opera and firefox
+    - Made the default music bot group configurable
 
 * **0.0.1-Alpha**:  
 - First public upload of the Web client